added pretty_print optional
